Formatos a los DataGridView y Reporte de venta por rango
Se le dio formato a las celdas de cada DataGridView en donde era
necesario, y se creo el reporte por rango de fechas, pues porqué se
tenia que hacer lo hacido, y aprovecho para poner esto porque no me
andas viendo :P

+++ GYM/Formularios/Lockers/frmLockers.cs
@@ -91,7 +91,7 @@ namespace GYM.Formularios
                     {
                         nombre = dr["nom_persona"].ToString();
                     }
-                    dgvLockers.Rows.Add(new object[] { dr["id"], dr["num"], nombre, tel, estado });
+                    dgvLockers.Rows.Add(new object[] { dr["id"], (int)dr["num"], nombre, tel, estado });
                 }
             }
             catch (MySqlException ex)
